refactor: streamline input relation registration and remove unused code
